requirements.txt
# Core Framework
flask>=2.3.3
flask-cors>=4.0.0
werkzeug>=2.3.7

# Discord Integration
discord.py>=2.3.2
aiohttp>=3.9.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.1
cryptography>=41.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# System Monitoring
psutil>=5.9.0

# Date/Time Handling
python-dateutil>=2.8.2

# Optional but Recommended
gunicorn>=21.2.0
